YAY localOffset is now 64 bits rather than 32

+++ src/bytecodegen.h
@@ -32,7 +32,7 @@ public:
     void gen(Node *node);
     void fixup();
 
-    void storeValue(Node *node, int32_t offset);
+    void storeValue(Node *node, int64_t offset);
 
     BytecodeGen();
 };
